--- allocation_pipeline/tests/test_feature.py
@@ -1,28 +1,0 @@
-import unittest
-import genomic_features
-
-
-class MyTestCase(unittest.TestCase):
-    def test_feature(self):
-        index = dict()
-        model = {"id": "ABCD00001", "version": 1, "children": [{"id": "ABCD00001_R0001",
-                                                                "children": [{"id": "ABCD00001_P0001", "version": 2}]}]}
-        feature = genomic_features.Feature(model, index)
-        self.assertEqual("ABCD00001", feature.source_id)
-
-    def test_gene(self):
-        index = dict()
-        model = {"source": "apollo", "id": "ABCD00001", "children": [{"id": "ABCD00001_R0001", "version": 2,
-                                                                "children": [{"id": "ABCD00001_P0001", "version": 2}]}]}
-        gene = genomic_features.ProteinCodingGene(model, index)
-        self.assertEqual("ABCD00001", gene.source_id)
-        self.assertEqual(1, len(gene.mrnas))
-        mrna = gene.mrnas[0]
-        self.assertEqual("ABCD00001_R0001", mrna.source_id)
-        cds = mrna.cds
-        self.assertEqual("ABCD00001_P0001", cds.source_id)
-
-
-
-if __name__ == '__main__':
-    unittest.main()
